requirements.txt
python-dotenv
requests
openai
jsonschema
gspread
pandas
pytest

